examples/python-examples/helloworld/requirements.txt
flask>=3.1.1
opentelemetry-distro>=0.45b0
opentelemetry-instrumentation-flask>=0.45b0
opentelemetry-instrumentation-requests>=0.45b0
opentelemetry-exporter-otlp>=1.24.0
pyroscope-io==0.8.8
pyroscope-otel>=0.1.0
typing-extensions>=4.12.2